extractor/requirements.txt
kafka-python
docling